SeqNode/requirements.txt
# ═══════════════════════════════════════════════════════════════
# SeqNode-OS — requirements.txt (versão pós-blueprint)
# ═══════════════════════════════════════════════════════════════

# ── API & Servidor (existente) ─────────────────────────────────
fastapi[standard]>=0.103.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
websockets>=12.0

# ── Clientes HTTP (existente) ──────────────────────────────────
httpx>=0.27.0
aiohttp>=3.9.0

# ── Modelos & Validação (existente) ───────────────────────────
pydantic>=2.3.0

# ── Orquestração DAG (existente) ──────────────────────────────
networkx>=3.1

# ── Config & Templates (existente) ────────────────────────────
pyyaml>=6.0.1
jinja2>=3.1.2

# ── I/O Assíncrono (existente) ────────────────────────────────
aiofiles>=23.2.1

# ── Base de Dados SQLite (existente) ─────────────────────────
aiosqlite>=0.19.0

# ── Métricas de Sistema (existente) ──────────────────────────
psutil>=5.9.5

# ── Fase 1 — MCP Server ──────────────────────────────────────
mcp>=1.0.0

# ── Fase 3 — Audit PDF ────────────────────────────────────────
# Instalar um dos dois (reportlab é mais completo, fpdf2 é mais leve):
reportlab>=4.0.0
# fpdf2>=2.7.0

# ── Fase 5 — Auth JWT ─────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.0.0
